pyproject.toml
[project]
name = "plastic-bottle-detection"
version = "0.1.0"
description = "project detect plastic bottle if it is with label or without label"
readme = "README.md"
requires-python = ">=3.9.13"

dependencies = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "onnxruntime-gpu==1.18.0",
    "ultralytics",
    "scikit-learn",
    "opencv-python",
    "pillow"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "flask>=3.1.2",
]
